Move Quirm into the engine
