+++ packages/voictents-and-estinants-engine/src/adapter/estinant-builder/output/outputHubblepupConditionalAppreffingeBuilder.ts
@@ -1,4 +1,4 @@
-import { Hubblepup } from '../../../core/engine-shell/quirm/hubblepup';
+import { Hubblepup } from '../../../core/engine-shell/hubblepup/hubblepup';
 import { GenericVoque } from '../../../core/engine/voque';
 import { Predicate } from '../../../utilities/predicate';
 import { SpreadN } from '../../../utilities/spreadN';
