+++ packages/voictents-and-estinants-engine/src/adapter/estinant-builder/right-input/rightInputHubblepupTupleAppreffingeBuilder2.ts
@@ -1,4 +1,4 @@
-import { GenericIndexedHubblepupTuple } from '../../../core/engine-shell/quirm/hubblepup';
+import { GenericIndexedHubblepupTuple } from '../../../core/engine-shell/hubblepup/hubblepup';
 import { GenericVoque } from '../../../core/engine/voque';
 import { Zorn2, ZornTuple2 } from '../../../utilities/semantic-types/zorn';
 import { SpreadN } from '../../../utilities/spreadN';
