--- packages/voictents-and-estinants-engine/src/core/engine-shell/quirm/quirm.ts
@@ -1,13 +1,0 @@
-import { Gepp } from '../voictent/gepp';
-import { Hubblepup } from './hubblepup';
-
-/**
- * One of the two programmable units of the Engine (see Estinant). It allows the Concrete Programmer to register
- * a Hubblepup to a Voictent via a Gepp.
- */
-export type Quirm = {
-  gepp: Gepp;
-  hubblepup: Hubblepup;
-};
-
-export type QuirmTuple = readonly Quirm[];
