+++ packages/voictents-and-estinants-engine/src/core/engine/inMemoryVoque.ts
@@ -1,5 +1,5 @@
 import { Merge } from 'type-fest';
-import { Hubblepup } from '../engine-shell/quirm/hubblepup';
+import { Hubblepup } from '../engine-shell/hubblepup/hubblepup';
 import { Gepp } from '../engine-shell/voictent/gepp';
 import { Voque } from './voque';
 import { AbstractSerializableIndexByName } from '../../core-programs/abstractSerializableVoictent';
